css/css-borders/corner-shape/corner-computed.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Borders and Box Decorations 4 Test: Computed values of 'corner'</title>
<link rel="author" title="Sebastian Zartner" href="mailto:sebastianzartner@gmail.com">
<link rel="help" href="https://drafts.csswg.org/css-borders-4/#corner-shorthand">
<meta name="assert" content="This test checks that the computed value of 'corner' is correct.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>

<div id="target"></div>

<script>
test_computed_value("corner", "0px round", "normal");
test_computed_value("corner", "normal");
test_computed_value("corner", "round 0rem", "normal");
test_computed_value("corner", "round 0%", "normal");
test_computed_value("corner", "4px 5% round", "4px 5% superellipse(1)");
test_computed_value("corner", "scoop 100%", "100% superellipse(-1)");
test_computed_value("corner", "30px superellipse(0.4)");
test_computed_value("corner", " 3% 5rem superellipse(0.4)", "3% 80px superellipse(0.4)");
test_computed_value("corner", "8px round / 10% bevel", "8px superellipse(1) / 10% superellipse(0)");
test_computed_value("corner", "10px 10% round / 10px 20% bevel / 16px round", "10px 10% superellipse(1) / 10px 20% superellipse(0) / 16px superellipse(1)");
test_computed_value("corner", "10px round / 33% superellipse(0.4) / superellipse(2) 6%/ 4px 12px scoop", "10px superellipse(1) / 33% superellipse(0.4) / 6% superellipse(2) / 4px 12px superellipse(-1)");
test_computed_value("corner", "4px round", "4px superellipse(1)");
test_computed_value("corner", "2% notch", "2% superellipse(-infinity)");
test_computed_value("corner", "4px round / 2% notch", "4px superellipse(1) / 2% superellipse(-infinity)");
test_computed_value("corner", "4px round / 2% round / 1em squircle", "4px superellipse(1) / 2% superellipse(1) / 16px superellipse(2)");
test_computed_value("corner", "4px round / normal / 1em round / 4% round", "4px superellipse(1) / normal / 16px superellipse(1) / 4% superellipse(1)");
test_computed_value("corner", "4px superellipse(3) / 2px superellipse(-0.3)");
test_computed_value("corner", "2% superellipse(1) / 2px superellipse(-1)");
test_computed_value("corner", "4px 2% squircle / round 2px", "4px 2% superellipse(2) / 2px superellipse(1)");
test_computed_value("corner", "4px bevel / superellipse(-3) 1em / 2px notch", "4px superellipse(0) / 16px superellipse(-3) / 2px superellipse(-infinity)");
test_computed_value("corner", "4px bevel / superellipse(-3) 1em / 2px notch / normal", "4px superellipse(0) / 16px superellipse(-3) / 2px superellipse(-infinity) / normal");
test_computed_value("corner", "4px superellipse(5) / notch 1em / 2px 13% round / bevel 10px", "4px superellipse(5) / 16px superellipse(-infinity) / 2px 13% superellipse(1) / 10px superellipse(0)");
</script>
