third_party/blink/web_tests/external/wpt/css/css-borders/corner-shape/corner-computed-expected.txt

This is a testharness.js-based test.
[FAIL] Property corner value '0px round'
  assert_true: corner doesn't seem to be supported in the computed style expected true got false
[FAIL] Property corner value 'normal'
  assert_true: corner doesn't seem to be supported in the computed style expected true got false
[FAIL] Property corner value 'round 0rem'
  assert_true: corner doesn't seem to be supported in the computed style expected true got false
[FAIL] Property corner value 'round 0%'
  assert_true: corner doesn't seem to be supported in the computed style expected true got false
[FAIL] Property corner value '4px 5% round'
  assert_true: corner doesn't seem to be supported in the computed style expected true got false
[FAIL] Property corner value 'scoop 100%'
  assert_true: corner doesn't seem to be supported in the computed style expected true got false
[FAIL] Property corner value '30px superellipse(0.4)'
  assert_true: corner doesn't seem to be supported in the computed style expected true got false
[FAIL] Property corner value ' 3% 5rem superellipse(0.4)'
  assert_true: corner doesn't seem to be supported in the computed style expected true got false
[FAIL] Property corner value '8px round / 10% bevel'
  assert_true: corner doesn't seem to be supported in the computed style expected true got false
[FAIL] Property corner value '10px 10% round / 10px 20% bevel / 16px round'
  assert_true: corner doesn't seem to be supported in the computed style expected true got false
[FAIL] Property corner value '10px round / 33% superellipse(0.4) / superellipse(2) 6%/ 4px 12px scoop'
  assert_true: corner doesn't seem to be supported in the computed style expected true got false
[FAIL] Property corner value '4px round'
  assert_true: corner doesn't seem to be supported in the computed style expected true got false
[FAIL] Property corner value '2% notch'
  assert_true: corner doesn't seem to be supported in the computed style expected true got false
[FAIL] Property corner value '4px round / 2% notch'
  assert_true: corner doesn't seem to be supported in the computed style expected true got false
[FAIL] Property corner value '4px round / 2% round / 1em squircle'
  assert_true: corner doesn't seem to be supported in the computed style expected true got false
[FAIL] Property corner value '4px round / normal / 1em round / 4% round'
  assert_true: corner doesn't seem to be supported in the computed style expected true got false
[FAIL] Property corner value '4px superellipse(3) / 2px superellipse(-0.3)'
  assert_true: corner doesn't seem to be supported in the computed style expected true got false
[FAIL] Property corner value '2% superellipse(1) / 2px superellipse(-1)'
  assert_true: corner doesn't seem to be supported in the computed style expected true got false
[FAIL] Property corner value '4px 2% squircle / round 2px'
  assert_true: corner doesn't seem to be supported in the computed style expected true got false
[FAIL] Property corner value '4px bevel / superellipse(-3) 1em / 2px notch'
  assert_true: corner doesn't seem to be supported in the computed style expected true got false
[FAIL] Property corner value '4px bevel / superellipse(-3) 1em / 2px notch / normal'
  assert_true: corner doesn't seem to be supported in the computed style expected true got false
[FAIL] Property corner value '4px superellipse(5) / notch 1em / 2px 13% round / bevel 10px'
  assert_true: corner doesn't seem to be supported in the computed style expected true got false
Harness: the test ran to completion.
